# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "wavynote",
            "email": "chelseafandev@gmail.com"
        }
    },
    "paths": {
        "/box/conversation": {
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "특정 대화방 삭제",
                "tags": [
                    "받은노트 페이지"
                ],
                "summary": "특정 대화방 삭제",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "uid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "conversation id",
                        "name": "cid",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/restapi.DefaultResponse"
                        }
                    },
                    "400": {
                        "description": "요청에 포함된 파라미터 값이 잘못된 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response400"
                        }
                    },
                    "401": {
                        "description": "인증에 실패한 경우이며, 실패 사유가 전달됩니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response401"
                        }
                    },
                    "404": {
                        "description": "요청한 리소스가 서버에 존재하지 않는 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response404"
                        }
                    },
                    "500": {
                        "description": "요청을 처리하는 과정에서 서버에 문제가 발생한 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response500"
                        }
                    }
                }
            }
        },
        "/box/conversationlist": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "노트를 주고 받는 대화방 목록 조회(최대 3개)",
                "tags": [
                    "받은노트 페이지"
                ],
                "summary": "노트를 주고 받는 대화방 목록 조회(최대 3개)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "조회된 대화방 목록 정보",
                        "schema": {
                            "$ref": "#/definitions/restapi.ConversationListResponse"
                        }
                    },
                    "400": {
                        "description": "요청에 포함된 파라미터 값이 잘못된 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response400"
                        }
                    },
                    "401": {
                        "description": "인증에 실패한 경우이며, 실패 사유가 전달됩니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response401"
                        }
                    },
                    "404": {
                        "description": "요청한 리소스가 서버에 존재하지 않는 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response404"
                        }
                    },
                    "500": {
                        "description": "요청을 처리하는 과정에서 서버에 문제가 발생한 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response500"
                        }
                    }
                }
            }
        },
        "/box/notelist": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "특정 친구와 주고받은 노트 목록",
                "tags": [
                    "받은노트 페이지"
                ],
                "summary": "특정 친구와 주고받은 노트 목록",
                "parameters": [
                    {
                        "type": "string",
                        "description": "conversation id",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/restapi.DefaultResponse"
                        }
                    },
                    "400": {
                        "description": "요청에 포함된 파라미터 값이 잘못된 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response400"
                        }
                    },
                    "401": {
                        "description": "인증에 실패한 경우이며, 실패 사유가 전달됩니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response401"
                        }
                    },
                    "404": {
                        "description": "요청한 리소스가 서버에 존재하지 않는 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response404"
                        }
                    },
                    "500": {
                        "description": "요청을 처리하는 과정에서 서버에 문제가 발생한 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response500"
                        }
                    }
                }
            }
        },
        "/box/show": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "특정 노트 조회",
                "tags": [
                    "받은노트 페이지"
                ],
                "summary": "특정 노트 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "conversation id",
                        "name": "cid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "note id",
                        "name": "nid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "uid",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "조회한 노트 정보",
                        "schema": {
                            "$ref": "#/definitions/restapi.NoteInfo"
                        }
                    },
                    "400": {
                        "description": "요청에 포함된 파라미터 값이 잘못된 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response400"
                        }
                    },
                    "401": {
                        "description": "인증에 실패한 경우이며, 실패 사유가 전달됩니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response401"
                        }
                    },
                    "404": {
                        "description": "요청한 리소스가 서버에 존재하지 않는 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response404"
                        }
                    },
                    "500": {
                        "description": "요청을 처리하는 과정에서 서버에 문제가 발생한 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response500"
                        }
                    }
                }
            }
        },
        "/main/folder": {
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "특정 폴더 이름 변경",
                "tags": [
                    "나의노트 페이지"
                ],
                "summary": "특정 폴더 이름 변경",
                "parameters": [
                    {
                        "description": "변경하고자 하는 폴더 정보",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/restapi.ChangeFolderNameRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/restapi.DefaultResponse"
                        }
                    },
                    "400": {
                        "description": "요청에 포함된 파라미터 값이 잘못된 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response400"
                        }
                    },
                    "401": {
                        "description": "인증에 실패한 경우이며, 실패 사유가 전달됩니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response401"
                        }
                    },
                    "404": {
                        "description": "요청한 리소스가 서버에 존재하지 않는 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response404"
                        }
                    },
                    "500": {
                        "description": "요청을 처리하는 과정에서 서버에 문제가 발생한 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response500"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "폴더 추가",
                "tags": [
                    "나의노트 페이지"
                ],
                "summary": "폴더 추가",
                "parameters": [
                    {
                        "description": "추가할 폴더 정보",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/restapi.AddFolderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/restapi.DefaultResponse"
                        }
                    },
                    "400": {
                        "description": "요청에 포함된 파라미터 값이 잘못된 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response400"
                        }
                    },
                    "401": {
                        "description": "인증에 실패한 경우이며, 실패 사유가 전달됩니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response401"
                        }
                    },
                    "404": {
                        "description": "요청한 리소스가 서버에 존재하지 않는 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response404"
                        }
                    },
                    "500": {
                        "description": "요청을 처리하는 과정에서 서버에 문제가 발생한 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response500"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "특정 폴더 삭제",
                "tags": [
                    "나의노트 페이지"
                ],
                "summary": "특정 폴더 삭제",
                "parameters": [
                    {
                        "description": "삭제할 폴더 정보",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/restapi.RemoveFolderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/restapi.DefaultResponse"
                        }
                    },
                    "400": {
                        "description": "요청에 포함된 파라미터 값이 잘못된 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response400"
                        }
                    },
                    "401": {
                        "description": "인증에 실패한 경우이며, 실패 사유가 전달됩니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response401"
                        }
                    },
                    "404": {
                        "description": "요청한 리소스가 서버에 존재하지 않는 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response404"
                        }
                    },
                    "500": {
                        "description": "요청을 처리하는 과정에서 서버에 문제가 발생한 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response500"
                        }
                    }
                }
            }
        },
        "/main/folderlist": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "존재하는 모든 폴더 목록 조회",
                "tags": [
                    "나의노트 페이지"
                ],
                "summary": "존재하는 모든 폴더 목록 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/restapi.FolderListResponse"
                        }
                    },
                    "400": {
                        "description": "요청에 포함된 파라미터 값이 잘못된 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response400"
                        }
                    },
                    "401": {
                        "description": "인증에 실패한 경우이며, 실패 사유가 전달됩니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response401"
                        }
                    },
                    "404": {
                        "description": "요청한 리소스가 서버에 존재하지 않는 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response404"
                        }
                    },
                    "500": {
                        "description": "요청을 처리하는 과정에서 서버에 문제가 발생한 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response500"
                        }
                    }
                }
            }
        },
        "/main/note": {
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "내가 쓴 특정 노트 삭제",
                "tags": [
                    "나의노트 페이지"
                ],
                "summary": "내가 쓴 특정 노트 삭제",
                "parameters": [
                    {
                        "description": "삭제할 노트 정보",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/restapi.RemoveNoteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/restapi.DefaultResponse"
                        }
                    },
                    "400": {
                        "description": "요청에 포함된 파라미터 값이 잘못된 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response400"
                        }
                    },
                    "401": {
                        "description": "인증에 실패한 경우이며, 실패 사유가 전달됩니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response401"
                        }
                    },
                    "404": {
                        "description": "요청한 리소스가 서버에 존재하지 않는 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response404"
                        }
                    },
                    "500": {
                        "description": "요청을 처리하는 과정에서 서버에 문제가 발생한 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response500"
                        }
                    }
                }
            }
        },
        "/main/notefolder": {
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "노트가 저장되는 폴더 변경",
                "tags": [
                    "나의노트 페이지"
                ],
                "summary": "노트가 저장되는 폴더 변경",
                "parameters": [
                    {
                        "description": "노트를 저장할 (변경할)폴더 정보",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/restapi.ChangeNoteFolderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/restapi.DefaultResponse"
                        }
                    },
                    "400": {
                        "description": "요청에 포함된 파라미터 값이 잘못된 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response400"
                        }
                    },
                    "401": {
                        "description": "인증에 실패한 경우이며, 실패 사유가 전달됩니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response401"
                        }
                    },
                    "404": {
                        "description": "요청한 리소스가 서버에 존재하지 않는 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response404"
                        }
                    },
                    "500": {
                        "description": "요청을 처리하는 과정에서 서버에 문제가 발생한 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response500"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "restapi.DefaultResponse": {
            "type": "object",
            "properties": {
                "msg": {
                    "description": "실패 시 반환하는 에러 메시지",
                    "type": "string"
                },
                "result": {
                    "description": "요청에 대한 처리 성공/실패 여부",
                    "type": "string",
                    "example": "true"
                }
            }
        },
...
        "restapi.Response400": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "msg": {
                    "type": "string",
                    "example": "Bad request"
                }
            }
        },
        "restapi.Response401": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 401
                },
                "msg": {
                    "type": "string",
                    "example": "Fail to parse Authorization header"
                }
            }
        },
        "restapi.Response404": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "msg": {
                    "type": "string",
                    "example": "Not exist the target resource in our server"
                }
            }
        },
        "restapi.Response500": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "msg": {
                    "type": "string",
                    "example": "Internal server error"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}